Move NewModuleTest and NewCriterionTest from test_nn.py to common_nn.py
